Get demos from demos folder instead of hardcoded list
